Let services report result status
In preperation to operate on an overall success or failed
state depending on the result of the individual services,
each service should have an easy way to report its result
state. The StatusReport class implements this in a way
that each service reports its state in a yaml file written
to the directory provided by get_status_report_directory().
As services are called by systemd and will run in parallel,
one file per service name is created to avoid potential
I/O conflicts on write. Issue #19