Handle missing credentials section as error
If no credentials section is specified in the config file
this should be reported as an error. That's because in the
image no user login will be allowed and therefore the
credentials setup from the config file is the only door
opener and mandatory for any subsequent work with the
system. This Fixes #29