Implement ssh key replacement to increase security
Instead of adding a private key from the config file to
the system the idea is to use the given private key only
temporary to inject a new key pair. This increases
security and opens the gate only until the temporary
private key is used in the key replacement process.

While the concept works it opens the problem that the
generated public key needs to be transported to the
customer endpoint. There is no process to establish
this in a clean which is the reason why the key
exchange code is currently not merged.

Maybe this change will never be merged but we'll keep
it open for later reference when the Azure large
instance environment would allow for processing
the idea here.

Related to Issue #42