Refactor reading of ssh private key
Instead of specifying a base64 encoded private ssh key
string as part of the key element in the config file,
we now just read a file path reference which is expected
to exist on the storage location from where the config
file was read. The referenced file is copied as the
ssh private key of the configured user. This sets the
config file free from any secret data and Fixes #58