Update schema semantic to specify ssh private key
The former ssh private key specification was based on
a custom file name and a target name for the key. The
given file name was copied and renamed as specified.
However the renaming seems superfluous and can be
avoided by specifying just the file name and copying
as such to the right place.