Allow reboot of the system via kexec
This patch set is two fold:

First it adds a reboot flag to the status report
and extends the cleanup service to allow for reboot of
the system via kexec if all services were successful
and at least one requests for reboot of the system in
its result report.

Second it extends the system setup service for the
crash kernel memory setup in a way that the high
value if not set via the config file is calculated
based on the amount of main memory. If the preconfigured
values have changed the system_setup service also
activates the reboot flag in its status report

This Fixes #64