Follow up fix for mount location checkup
Provides a umount_config_source method in the Defaults
space to allow to perform an umount which succeeds even
if the mountpoint is busy. The umount is performed as
Lazy unmount and detaches the filesystem from the file
hierarchy now, and clean up all references to this
filesystem as soon as it is not busy anymore.
This Fixes #76