Fix inconsistent cleanup service
The introduction of the pre/post scripts in the package which
runs systemd macros to delete the services causes a core dump
on zypper and prevents the cleanup service from actually
uninstalling the package. Thus I suggest to delete those
pre/post actions from the package and leave this in the hand
of the cleanup service using the reset-failed command from
systemd after uninstall of the package has been performed.
This references Issue #88