Added workload status flag file
At the time of the cleanup service a file named:
workload_success_is_true or workload_success_is_false
will be written on the storage location the config file
was read from. At the time this file appears it's also
safe to release the config file storage location from
the system. This Fixes #93