Write workload log file
In case of a deployment error, a log file in addition to the
status flag files is written which contains the systemd
service log information from all services ran so far.
The log file is written to the storage location from which
the yaml config file was read from. It is expected that
this location is writable and offers enough space to store
the logfile. In case of an error writing the log we will
treat this as "bad luck" and continue with the cleanup.
This Fixes #96