Fixed group processing in setup_ssh_authorization
The change in the schema to specify group names/id was not
handled in the code processing the setup of the ssh
authorization. As there was a group key before and after the
change in the schema but with a different result and used
only as parameters to a mocked grp.getgrnam() system call
the unit tests did not uncover the issue. This Fixes #98