Support MTU setup for standard and vlan interfaces
Enhanced the config schema to allow the networking attributes
mtu and vlan_mtu as optional arguments to configure the maximum
transfer unit for the interface and/or its virtual type. In
addition add a differentiator for the VeryLarge instance types
in generation 3 and its successor. For Gen3 VeryLarge instance
types the network setup was not yet defined because we haven't
started to support this target. However the successor of the
VeryLarge instance type is similar in the infrastructure and
boot compared to the LargeInstance type and can be covered
by the network service in the same way. This Fixes #105
and Fixes bsc#1120522