Fixed network service vlan setup
The vlan setup used the interface name in its ifcfg config
file in the following format: ifcfg-<interface>.<vlan_id>
However if the interface name is based on the MAC address
that leads to a name which is longer than the fixed defined
value from /usr/include/linux/if.h: #define IFNAMSIZ 16
and will make the network setup to fail. Therefore this
patch changes the config file for the vlan setup to follow
the format: ifcfg-vlan<vlan_id> which avoids that size
limit. This Fixes #115