Extend storage service dependencies
The storage service can be used for remote storage like NFS
storage to be attached to the machine. This requires the network
to be online. Having the network only configured is not enough
it must also be online. Thus the storage service unit is extended
to wait for the network-online.target