Added more logging to the process
Add a log file /var/log/azure-li-services.log which adds
logging information from the service process. Usually error
log information is present on the systemd level but for
checking the process, it's calls and potential further information
it's also useful to have a processing log file. The log file
will be created on the host and gets also copied to the config
lun in the same way as the systemd workload log