Consolidate all image descriptions in git
Instead of maintaining image descriptions in obs we want to
maintain them in git. With this change only a service and
multibuild configuration applies in obs but the data to
build the image will live in git. This allows for real
development and review regarding changes to the kiwi image
descriptions.