Right sequence saptune
One of the issues is that `saptune` is a different tool that supersedes
`sapconf`. Then the `saptune daemon restart` command will always
overwrite the profile with `saptune`. Two different tools that can't be
mixed. Only one should be used.

In case of SLES (not SLES for SAP), the sequence should be
For SLES 12
```
tuned-adm profile sap-hana
systemctl enable --now sapconf.service
```
and for SLES15
```
tuned-adm profile sapconf
systemctl enable --now sapconf.service
```
For SLES for SAP, the sequence is the same for 12 and 15:
```
saptune daemon start
saptune solution apply HANA
```

This Fixes #172