Add obs artificats
This pipeline job creates the static image files as they
would be needed in the obs system if a source service
is not allowed. This is usally the case for production
submissions. Thus the data on the artificats link in
gitlab now offers this data and can just be used to
drive these type of submissions