Fixed High value calculation in kdump setup
The High value gets calculated to match the requirements
of the systems used at Microsoft. The result of the testing
leads to the empiric formular dump-high * max(1, mem-TB) + LUN/2
We know that this is not matching the various recommendations
out on the web but this was reported to be working on the
high mem systems used at Microsoft. We rate functionality
over recommendations. This Fixes #203