Fix printk timing issue
This change adds 'loglevel=3' to the kernel command line.
This is needed as a workaround to handing console messages
as a kthread. In certain cases there colud be a process block
while printing console messages. One viable worksround is to
adjust the loglevel to '3' to only show messages for "error,
critical, alert, and emergency."