Setup link name policy
Setup name policy from file in /usr/lib/systemd/network. Given a
correct lexical name to be evaluated by systemd before the udev
provided 99-default.link file, allows to setup the link name policy
as we want it. The implementation here is the preferred solution
according to the systemd maintainers and replaces the former
solution using a config file in /etc/systemd/network.
Related to bsc#1176738