Explicitly activate multipath lookup
All LI/VLI images runs from a SAN storage via redundant
storage paths managed by multipath. The implicit check
for eventually existing multipath devices uncovered a
potential race condition if the multipath maps appears
too late. This is related to Issue #255. As a consequence
a fix in kiwi was made OSInside/kiwi#1641 which checks
for multipath maps until they appear or fail if they
don't appear after a timeout. This change requires an
image that is designed to run in a multipath environment
to explicitly request it, which is done in this PR